Store and buying implementation completed
Finished implementing the ability for the player to buy equipment at the store at the start of the beginning. Store class and methods are implemented and buying is also possible. Created a test weapon called "spear" in core_mechanics.py but that will not necessarily be the final Spear for the game.