New buying function that handles invalid inputs
Buying implementation that is smoother and less clunky. Players can directly type "buy" or "refund" followed by the name of the item they want to buy. Used the string split method to do this, along with a try and except structure to handle cases of an invalid input that is only one word or multiple words.