Implemented Wall decisions
Wrote the story up to and player encountering the misty wall. Wrote function to facilitate the decision making for the player's actions at the wall. Created use_backpack method for the Player class that returns an item of their choice from their backpack. Edited the repr and str methods for the classes to include the class type of each instance. Next steps are to figure out the battle mechanics for the first fight.