Place classes implemented
Moved the buying function to the core_mechanics file for sake of consistency. Created the Place class and it's randomization functionality. Places have class attributes containing the possible size, enemies, and events of the instance. Instances are randomized based on their type, which is default random. The fill method is what populates the Place instance randomly, based on its type. Different types on places will have their own subclass with their own lists of possibilities.