Battle mechanics fully implemented
All of the battle mechanics are fully implemented. Battles can now be played when the player ends up on an Enemy or Tutorial Place. Turn based combat is now implemented, where the player has the options to move, attack, or use an item in their backpack. Rudimentary AI for the enemy turns are also implemented. Implemented the Weapon class's action method to allow for the weapons to deal damage. Created an injure and remove method to allow for damaging Entities and their removal if their health hits 0. Changed the use_backpack method to allow the player to directly type in the action they want to do and the item they want to do the action on, similar to the buying function's implementation. Players can now remove items from their backpack, which are lost forever unless they find a new instance of it. Enemies have combat dialogue and death dialogue. Attempted to polish the text display so it looks nicer but needs some work with the spacing.