Backwards movement
Following suggestions from my girlfriend, modified the player's movement method to allow the player to go backwards if they type in a negative number of steps. Also modified the method so that if the player is immediately blocked, they can still try their turn again. However, if they are able to move at least one step in the chosen direction, they will have moved and used their turn. Also modified the battle lines for enemies so that it occurs after the turn header displays and had the chances for the dialogue displaying as a class attribute.