Event classes and Three Perks Event
Implemented the Event class and a working version of the Three Perks event that allows the player to pick a bonus of acquire. Also implemented a Bow class and a Lyre class so that two of the perks will function correctly. Did a single quick test and the current implementation works so far. For the Equipment subclasses, changed the damage, healing, and range calculations in the action methods to include the base stat, a subclass stat, and any bonuses tied to the player, such as an overall damage increase. All of these changes were to allow for the buff and debuff system that comes along with the Three Perks event. Changed the order of the Event section and the Place section so that the Place section is defined last, allowing it to take full advantage of all the previously defined classes.