New buff and debuff implementation
New system to calculate bonuses for the various classes in game by giving each class an effective version of a stat that is regularly recalcuated with the recalculate method. This method recalculates the value of these effective stats by summing up all the related bonuses. The Player class's recalculate method does this for the player and all items in the backpack. Stats are recalculated when a stat is changed in the method of an instance or at the start of every turn. Now, player bonuses are universal, class specific bonuses only affect instances of a certain class, and place bonuses only matter if the player is on a certain type of place. The Three Perks play method was changed so that it is shorter and more segmented, with the code for each god choice being a its own method. Tested the functionality of the code and it works. Added a turn counter for Place instances, which may be usable for turn based bonuses. Made a test bow weapon to test the ranged combat in the game. It is working!