Rifle class, Legendary Item event
Fully implemented the rifle class of ranged weapons that have an accuacy stat.  Legendary item event is also written, both the code and the text behind it. Currently only have the Ross Rifle item. Modified the text for some methods so that it looks better. Showing the backpack now also shows the player's weight. Players have the option to use their backpack after finishing a fight or event before they continue. Fixed an issue where the player could use a weapon or healing tool in this moment, causing them to take a turn that they cannot get out of. Added some conditional statements to prevent this. Place visualization now separates enemies with a / mark to differentiate them. Added __init__ methods for subclasses that didn't have it for consistancy. Jungle Places now have a range deteriment of 1 unit. Fixed an issue with the Place type randomization that allowed a place to be a tutorial place, which shouldn't happen.