Feral Monkeys, Sorting method, Show inventory changes
Implemented the Feral Monkeys enemy class and implemented it into the enemy constructor and Jungle Place class. The sorting method for the Player class was created to allow item sorting for non combat situations, preventing the player from using a weapon or a healing item at max health. These two actions would otherwise break the game. The onwards function and the other sections of code that give the player a chance to fix up their backpack are now changed to use this method. The show inventory method and the show backpack methods have been slightly alerted in their implementation so that there is less code. All place instances have a default Event instance, which has a play method that gives the player a free space and lets them just organize their backpack before going on. This allows for the game to slowly take out already done Event instances in the possible events list of each Place subclass so that events do not occur twice. If there is no event left, an Event place will default to the aforementioned base Event instance. Some additional print lines were used to fix the text formatting. Equipment classes and subclasses are given a class name attribute to help with the new inventory implementation.