Merchant, Ghostly Vision event, Place add_enemy, Player eff_max_health
Finished some dialogue for the Ghostly Vision event and finished the implementation of the Merchant event. Both events currently untested! Added an add_enemy method to the Place class so that in the future, enemies can spawn in during a battle. Added a player eff_max_health attribute so that some places can boost the player's max health. Changed the Healing Equipment action method so that it uses this new effective max health instead of the base stat. Added a recalculation to the Mars bonus so that the code recalculates the new player stats.