Giant Snake, Store class inventory fix, Enemy backwards move
Implemented the Giant Snake boss for the jungle area and wrote the dialogue for it as well. Currently working as intended. Store class had an issue where adding an item to one store added it to all stores. Fixed this issue by removing the item argument for the __init__ method. Enemy move method has an option back argument which is a booleon saying if the enemy is going to move back or not. Enemies can now move backwards. Made some placeholder new items to test out.