Alpha Harpy class structure, New text structure, New enemy constructor implementation
New implementation of the enemy constructor so that it takes in a string of the name of the class and creates a new instance by evaluating the string plus the call parentheses. Lost a lot of progress due to GitHub's terrible UI but managed to create a structure for the new text and methods needed to be written for the Alpha Harpy enemy class.