Lab introduction written, Structure for enemy classes written
Wrote the story up to the introduction for entering the Machine Labs. Implemented the Place class for the labs as well, including all the bonuses for it. Wrote the basic structure for the enemy classes for this area but have yet to implement any of them. Quoted out portions of the game in the main file to allow for quicker testing and readded a line of code for the Cliffs that was accidentally deleted.