Rewriting story, implementing first few in game choices
Decided to rework the entire game's theme to be that of a dystopian prison escape. This requires almost everything in the game's story and main to be changed for the new story. So far, implemented the first few choices of the game up to where the player is introduced to the guards. Legionary class has been changed to a Baton Guard class but the rest of the classes are unchanged for now. Deleted the City Navigation and Dodge events since those are not needed anymore. Removed place bonuses from all places and changed recalculate methods to match this. Cell place class created for the first fight. Player now has a permanent fist weapon. Parts of the story have been rewritten and the choices function has been made to allow for the player to make choices. Working as intended thus far. Deleted all the other Place subclasses aside from the Machine Labs, which will be used. Deleted the main file's code and started from the beginning.