Escape sequence implemented
Changed all the usage of the word 'backpack' to 'inventory' and the backpack is now called the inventory. Most events were deleted and the Lyre equipment class was removed. The price attribute for all equipment is now removed and the str methods have been updated to reflect this. Player wallet attribute is removed but the Store class is still in the core mechanics file for now in the event it can be reused for something. The escaping function is implemented in the main file to allow the player to make decisions if they choose to not fight the guard right away. Choices have been tested and everything is work as intended. Story is written up to the player killing the first guard.