Looting mechanic implemented
Implemented the looting mechanic by reusing the methods from the Store class's buying functionality. Now the player can pick up loot from enemies after the battle is over using a very similar manner to buying things. Enemies now have a drop chance for loot and a list of possible loot that is the code to create an instance of equipment in string format. Once selected, the string is evaluated to create a new instance of the loot, allowing generation of new loot. Dropping loot can be turned off by using the can drop class attribute. Fixed a bug in the use weapons method for the player that crashes the program if the player selects a value that is out of index. Continued the story up to the player fighting the second guard.