Volk boss fight finished, Tweaked stats of certain items
Tweaked the stats of certain items to make them more balanced and reasonable for gameplay. Increased the item drop rate for all enemies to help in combat. Completed the story up to the Volk boss fight, including the dialogue, story, and game play of the boss and the arena. Not all new changes have been tested.